# docs/rebuild/openapi-baseline.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Tradier Bot API",
  "description": "Standalone backend for three worlds on one process: the Tradier options desk (balance, chain preview, managed positions with TP/SL, the opening-range and A/B auto-traders, the HOT scan and the options-flow board), the 36 Trade Desk that fronts it on a phone, and the Bot Station that launches, supervises and settles the vendored Kalshi bots (btc15, btc60, sports, parlay, gold15, silver15, oil15). Plus the super_research signal engines and the SPY/QQQ/SPX level watcher they read. SQLite persistence, per-user credential folders, paper-only by default.",
  "version": "1.0.0"
 },
 "paths": {
...
  "/api/v1/super/off": {
   "post": {
    "tags": [
     "super-research"
    ],
    "summary": "Super Off",
    "description": "Stop supervisors (all, or one category). The legacy stack has no stop\nendpoint \u2014 this replaces `python bots.py stop`.",
    "operationId": "superOff",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "anyOf": [
         {
          "$ref": "#/components/schemas/SuperStopRequest"
         },
         {
          "type": "null"
         }
        ],
        "title": "Payload"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "type": "object",
         "title": "Response Superoff"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/super/config": {
   "get": {
    "tags": [
     "super-research"
    ],
    "summary": "Get Config",
    "operationId": "getSuperConfig",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "type": "object",
         "title": "Response Getsuperconfig"
        }
       }
      }
     }
    }
   },
   "post": {
    "tags": [
     "super-research"
    ],
    "summary": "Set Config",
    "operationId": "setSuperConfig",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/SuperConfigUpdate"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "type": "object",
         "title": "Response Setsuperconfig"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/super/quote/{ticker}": {
   "get": {
    "tags": [
     "super-research"
    ],
    "summary": "Get Quote",
    "description": "Live price + classic pivot points for a desk ticker (yfinance,\nkeyless, cached 60s). Includes the TradingView chart URL.",
    "operationId": "getTickerQuote",
    "parameters": [
     {
      "name": "ticker",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Ticker"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Gettickerquote"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/super/gex0dte/refresh": {
   "post": {
    "tags": [
     "super-research"
    ],
    "summary": "Refresh Gex0Dte",
    "operationId": "refreshGex0dte",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/Gex0dteRefresh"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "type": "object",
         "title": "Response Refreshgex0Dte"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/super/engine-pct": {
   "get": {
    "tags": [
     "super-research"
    ],
    "summary": "Get Engine Pct",
    "description": "Per-category TP/SL race targets, read from each ticker's config.py.\n\n``mixed: true`` means the tickers in that category disagree \u2014 reported\nrather than averaged, because it usually means a folder was hand-edited.",
    "operationId": "getEnginePct",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "type": "object",
         "title": "Response Getenginepct"
        }
       }
      }
     }
    }
   },
   "post": {
    "tags": [
     "super-research"
    ],
    "summary": "Set Engine Pct",
    "description": "Rewrite TP_PCT/SL_PCT for every ticker in a category.\n\nTakes effect on the next scan cycle. The engine-score cache is\nfingerprinted on tp/sl so it re-scores itself; the B-book candidate set is\nnot re-discovered, which the response says explicitly.",
    "operationId": "setEnginePct",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/EnginePctUpdate"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "type": "object",
         "title": "Response Setenginepct"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/super/engine-gates": {
   "get": {
    "tags": [
     "super-research"
    ],
    "summary": "Get Engine Gates",
    "description": "A-book / B-book tp-before-sl admission gates (desk-wide).\n\nThese are module-level constants in engine_common.py shared by every\nticker, so there is one pair of numbers, not one per category.",
    "operationId": "getEngineGates",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "type": "object",
         "title": "Response Getenginegates"
        }
       }
      }
     }
    }
   },
   "post": {
    "tags": [
     "super-research"
    ],
    "summary": "Set Engine Gates",
    "description": "Set the A/B admission gates. Applies on the next scan; no re-scoring.",
    "operationId": "setEngineGates",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/EngineGatesUpdate"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "type": "object",
         "title": "Response Setenginegates"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/super/gex0dte/heartbeat": {
   "post": {
    "tags": [
     "super-research"
    ],
    "summary": "Gex0Dte Heartbeat",
    "description": "Record that a push cycle happened, whatever its outcome.\n\nStrictly separate from the data path: it never fetches the vendor, never\nstores a snapshot and never fills an hour slot. Routing liveness through\n/refresh would make the SERVER call getgamma on every failed cycle \u2014 the\none thing this design does not do \u2014 and would stamp a stalled feed as\nfresh.",
    "operationId": "gex0dtePusherHeartbeat",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/HeartbeatIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "type": "object",
         "title": "Response Gex0Dtepusherheartbeat"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/super/tickers": {
   "post": {
    "tags": [
     "super-research"
    ],
    "summary": "Add Ticker",
    "description": "Scaffold, register and bootstrap a new ticker in a category.\n\nReturns as soon as the folder and registry entry exist; the B-book build\n(60 days of bars) runs detached, so poll ``/tickers/{id}/status`` for it.\nThe supervisors pick the ticker up on their next cycle \u2014 they re-read the\nregistry \u2014 and unconditionally after a restart.",
    "operationId": "addSuperTicker",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/AddTickerRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "type": "object",
         "title": "Response Addsuperticker"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/super/tickers/{ticker_id}/status": {
   "get": {
    "tags": [
     "super-research"
    ],
    "summary": "Ticker Status",
    "description": "Whether the background bootstrap has produced a playbook yet.",
    "operationId": "getSuperTickerStatus",
    "parameters": [
     {
      "name": "ticker_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Ticker Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Getsupertickerstatus"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/super/gex0dte/history": {
   "get": {
    "tags": [
     "super-research"
    ],
    "summary": "Get Gex0Dte History",
    "description": "Hourly SPY 0DTE net gamma for one trading day, 08:00-16:00 CST.\n\nAll nine hours come back every time; ones never captured read 0 and carry\n``captured: false``, so a quiet hour is distinguishable from a flat one.",
    "operationId": "getGex0dteHistory",
    "parameters": [
     {
      "name": "date",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "description": "YYYY-MM-DD (CST); omit for today",
       "title": "Date"
      },
      "description": "YYYY-MM-DD (CST); omit for today"
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Getgex0Dtehistory"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "Gex0dteRefresh": {
    "properties": {
     "payload": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Payload"
     },
     "ticker": {
      "type": "string",
      "title": "Ticker",
      "default": "SPY"
     },
     "client": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Client"
     }
    },
    "type": "object",
    "title": "Gex0dteRefresh",
    "description": "Give it a chain captured from a getgamma.io tab, or nothing to make the\nserver try the vendor directly. No credentials either way \u2014 the endpoint\nneeds none; its edge simply refuses non-browser clients."
   },
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "SuperConfigUpdate": {
    "properties": {
     "enabled": {
      "additionalProperties": {
       "type": "boolean"
      },
      "type": "object",
      "title": "Enabled",
      "description": "tickerId -> enabled flag, matched across all categories",
      "examples": [
       {
        "btc": false,
        "spy": true
       }
      ]
     }
    },
    "type": "object",
    "required": [
     "enabled"
    ],
    "title": "SuperConfigUpdate"
   },
...
   "SuperStopRequest": {
    "properties": {
     "category": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Category",
      "description": "etf | stock | crypto; omit to stop all"
     }
    },
    "type": "object",
    "title": "SuperStopRequest"
   },
...
  }
 }
}